requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
argparse==1.4.0
array-record==0.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
automat==22.10.0
blessed==1.20.0
blinker==1.7.0
buildtools==1.0.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
constantly==23.10.4
contourpy==1.2.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
dill==0.3.7
dm-tree==0.1.8
docker==6.1.3
docopt==0.6.2
docstring-parser==0.16
einops==0.7.0
entrypoints==0.4
etils==1.7.0
filelock==3.13.1
flask==3.0.0
flatbuffers==24.3.25
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.40
google-pasta==0.2.0
gpustat==1.1.1
greenlet==3.0.2
grpcio==1.65.1
gunicorn==21.2.0
h5py==3.11.0
huggingface-hub==0.19.4
hyperlink==21.0.0
idna==3.6
immutabledict==4.2.0
importlib-metadata==7.0.0
importlib-resources==6.4.0
incremental==22.10.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.3.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
mako==1.3.0
markdown==3.5.1
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.4.0
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
namex==0.0.8
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
nvidia-ml-py==12.555.43
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.12.1
orderedmultidict==1.0.1
packaging==23.2
pandas==2.1.3
pillow==10.1.0
platformdirs==4.1.0
prettytable==3.9.0
promise==2.3
protobuf==3.20.3
psutil==6.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pygments==2.18.0
pyjwt==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytz==2023.3.post1
pyyaml==6.0.1
querystring-parser==1.2.4
redo==2.0.4
regex==2023.10.3
requests==2.31.0
rich==13.7.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
simple-parsing==0.1.5
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sqlalchemy==2.0.23
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-datasets==4.9.6
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.15.0
termcolor==2.4.0
tfds-nightly==4.4.0.dev202201230107
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
transformers==4.35.2
triton==2.1.0
twisted==24.3.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
visualize==0.5.1
wcwidth==0.2.12
websocket-client==1.7.0
werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope-interface==6.3